// cinescope/src/app/core/dtos/people/person-movie-credits.dto.ts
export interface TmdbPersonMovieCastDto {
    adult: boolean;
    backdrop_path: string | null;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
    character: string;
    credit_id: string;
    order: number;
}

export interface TmdbPersonMovieCrewDto {
    adult: boolean;
    backdrop_path: string | null;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
    credit_id: string;
    department: string;
    job: string;
}

export interface TmdbPersonMovieCreditsResponseDto {
    id: number;
    cast: TmdbPersonMovieCastDto[];
    crew: TmdbPersonMovieCrewDto[];
}